// client/src/utilities/Types.ts
// Reg. Form State Type
export interface RegFormStateType {
  firstName: string;
  middlename: string;
  surname: string;
  email: string;
  password: string;
  phoneNumber: string;
  confirmPwd: string;
  address: string;
  role: string;
}

// Login Form State Type
export interface LoginFormStateType {
  email: string;
  password: string;
}